# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ArtistHub Social Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /artists:
    get:
      operationId: getArtists
      tags: [artists]
      summary: List/search artists
      parameters:
        - name: location
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: tags
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Artists list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistsResponse"
  /artists/{userId}:
    get:
      operationId: getArtistProfile
      tags: [artists]
      summary: Get artist profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /artists/{userId}/update:
    post:
      operationId: updateArtistProfile
      tags: [artists]
      summary: Update artist profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateArtistRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
  /artists/{userId}/gallery:
    post:
      operationId: addGalleryItem
      tags: [artists]
      summary: Add a gallery item to artist profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemRequest"
      responses:
        "201":
          description: Added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /artists/{userId}/gallery/{itemId}:
    delete:
      operationId: deleteGalleryItem
      tags: [artists]
      summary: Delete a gallery item
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /events:
    get:
      operationId: getEvents
      tags: [events]
      summary: List events
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: city
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{eventId}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
    RegisterRequest:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required: [username, email, password]
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        message:
          type: string
      required: [user]
...
    User:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        profileType:
          type: string
          enum: [user, artist]
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        followerCount:
          type: integer
        friendCount:
          type: integer
        followingCount:
          type: integer
        postCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        bannerUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        work:
          type: string
          nullable: true
        school:
          type: string
          nullable: true
        about:
          type: string
          nullable: true
        interests:
          type: array
          items:
            type: string
        hasArtistPage:
          type: boolean
        accentColor:
          type: string
          nullable: true
        themeName:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        onboardingStep:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/LinkItem"
        featuredContent:
          type: string
          nullable: true
      required: [id, username, email, profileType, isAdmin, isBanned, followerCount, friendCount, followingCount, postCount, createdAt]
    LinkItem:
      type: object
      properties:
        label:
          type: string
        url:
          type: string
      required: [label, url]
    ProfileCustomField:
      type: object
      properties:
        label:
          type: string
        value:
          type: string
      required: [label, value]
    UserSummary:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        artistDisplayName:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        profileType:
          type: string
          enum: [user, artist]
        followerCount:
          type: integer
        friendCount:
          type: integer
        bannerUrl:
          type: string
          nullable: true
        location:
          type: string
          nullable: true
        city:
          type: string
          nullable: true
        age:
          type: integer
          nullable: true
        work:
          type: string
          nullable: true
        school:
          type: string
          nullable: true
        about:
          type: string
          nullable: true
        interests:
          type: array
          items:
            type: string
        hasArtistPage:
          type: boolean
        accentColor:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        onboardingStep:
          type: string
        links:
          type: array
          items:
            $ref: "#/components/schemas/LinkItem"
        category:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        hasBlockedUser:
          type: boolean
        isBlockedByUser:
          type: boolean
      required: [id, username, profileType, followerCount, friendCount, hasBlockedUser, isBlockedByUser]
...
    Post:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        actorSurface:
          type: string
          enum: [personal, artist]
        content:
          type: string
        imageUrl:
          type: string
          nullable: true
        visibility:
          type: string
          enum: [public, friends, private]
        repostOfPostId:
          type: integer
          nullable: true
        likeCount:
          type: integer
        isLiked:
          type: boolean
        reactionCounts:
          $ref: "#/components/schemas/ReactionCounts"
        totalReactionCount:
          type: integer
        currentUserReaction:
          type: string
          nullable: true
          enum: [like, heart, wow, angry]
        repostCount:
          type: integer
        commentCount:
          type: integer
        comments:
          type: array
          items:
            $ref: "#/components/schemas/PostComment"
        originalPost:
          allOf:
            - $ref: "#/components/schemas/Post"
          nullable: true
        author:
          $ref: "#/components/schemas/UserSummary"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        media:
          type: array
          items:
            $ref: "#/components/schemas/PostMedia"
      required: [id, userId, actorSurface, content, visibility, likeCount, isLiked, reactionCounts, totalReactionCount, repostCount, commentCount, comments, createdAt]
...
    ReactionCounts:
      type: object
      properties:
        like:
          type: integer
        heart:
          type: integer
        wow:
          type: integer
        angry:
          type: integer
      required: [like, heart, wow, angry]
...
    ArtistProfile:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        displayName:
          type: string
          nullable: true
        category:
          type: string
        location:
          type: string
          nullable: true
        tagline:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
        influences:
          type: string
          nullable: true
        availabilityStatus:
          type: string
          nullable: true
        pronouns:
          type: string
          nullable: true
        yearsActive:
          type: string
          nullable: true
        representedBy:
          type: string
          nullable: true
        openForCommissions:
          type: boolean
        touring:
          type: boolean
        acceptsCollaborations:
          type: boolean
        customFields:
          type: array
          items:
            $ref: "#/components/schemas/ProfileCustomField"
        bookingEmail:
          type: string
          nullable: true
        gallery:
          type: array
          items:
            $ref: "#/components/schemas/GalleryItem"
        user:
          $ref: "#/components/schemas/UserSummary"
        primaryActionType:
          type: string
        primaryActionLabel:
          type: string
        primaryActionUrl:
          type: string
          nullable: true
        featuredTitle:
          type: string
          nullable: true
        featuredDescription:
          type: string
          nullable: true
        featuredUrl:
          type: string
          nullable: true
        featuredType:
          type: string
        moodPreset:
          type: string
        layoutTemplate:
          type: string
        fontPreset:
          type: string
        enabledModules:
          type: array
          items:
            type: string
        moduleOrder:
          type: array
          items:
            type: string
        pinnedPost:
          $ref: "#/components/schemas/Post"
          nullable: true
      required: [id, userId, category, tags, gallery, user, primaryActionType, primaryActionLabel, openForCommissions, touring, acceptsCollaborations, customFields, featuredType, moodPreset, layoutTemplate, fontPreset, enabledModules, moduleOrder]
    UpdateArtistRequest:
      type: object
      properties:
        displayName:
          type: string
          nullable: true
        category:
          type: string
        location:
          type: string
          nullable: true
        tagline:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        bio:
          type: string
          nullable: true
        influences:
          type: string
          nullable: true
        availabilityStatus:
          type: string
          nullable: true
        pronouns:
          type: string
          nullable: true
        yearsActive:
          type: string
          nullable: true
        representedBy:
          type: string
          nullable: true
        openForCommissions:
          type: boolean
        touring:
          type: boolean
        acceptsCollaborations:
          type: boolean
        customFields:
          type: array
          items:
            $ref: "#/components/schemas/ProfileCustomField"
        bookingEmail:
          type: string
          nullable: true
        primaryActionType:
          type: string
        primaryActionLabel:
          type: string
        primaryActionUrl:
          type: string
          nullable: true
        featuredTitle:
          type: string
          nullable: true
        featuredDescription:
          type: string
          nullable: true
        featuredUrl:
          type: string
          nullable: true
        featuredType:
          type: string
        moodPreset:
          type: string
        layoutTemplate:
          type: string
        fontPreset:
          type: string
        enabledModules:
          type: array
          items:
            type: string
        moduleOrder:
          type: array
          items:
            type: string
        pinnedPostId:
          type: integer
          nullable: true
    GalleryItem:
      type: object
      properties:
        id:
          type: integer
        artistId:
          type: integer
        type:
          type: string
          enum: [image, video, audio]
        url:
          type: string
        caption:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, artistId, type, url, createdAt]
    GalleryItemRequest:
      type: object
      properties:
        type:
          type: string
          enum: [image, video, audio]
        url:
          type: string
        caption:
          type: string
          nullable: true
      required: [type, url]
...
    ArtistsResponse:
      type: object
      properties:
        artists:
          type: array
          items:
            $ref: "#/components/schemas/ArtistProfile"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [artists, total, page, totalPages]
...
    Event:
      type: object
      properties:
        id:
          type: integer
        hostUserId:
          type: integer
        title:
          type: string
        description:
          type: string
        startsAt:
          type: string
          format: date-time
        location:
          type: string
        city:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        lineupTags:
          type: array
          items:
            type: string
        visibility:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        host:
          $ref: "#/components/schemas/UserSummary"
          nullable: true
        artists:
          type: array
          items:
            $ref: "#/components/schemas/UserSummary"
      required: [id, hostUserId, title, description, startsAt, location, lineupTags, visibility, createdAt, updatedAt, artists]
    CreateEventRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startsAt:
          type: string
          format: date-time
        location:
          type: string
        city:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        lineupArtistIds:
          type: array
          items:
            type: integer
        lineupTags:
          type: array
          items:
            type: string
      required: [title, description, startsAt, location]
...